[sw/lib] move CSR definitions to separate file
